+++ app/src/Controller/ProjectController.php
@@ -18,11 +18,16 @@ use Symfony\Component\Routing\Annotation\Route;
  */
 class ProjectController extends AbstractController
 {
+
     /**
      * @Route("/", name="project_index", methods={"GET"})
      */
     public function index(ProjectRepository $projectRepository): Response
     {
+        if (!$this->getUser()) {
+            return $this->redirectToRoute('root');
+        }
+        
         return $this->render('project/index.html.twig', ['projects' => $projectRepository->findAll()]);
     }
 
@@ -31,6 +36,10 @@ class ProjectController extends AbstractController
      */
     public function new(Request $request): Response
     {
+        if (!$this->getUser()) {
+            return $this->redirectToRoute('root');
+        }
+
         $project = new Project();
         $form = $this->createForm(ProjectType::class, $project);
         $form->handleRequest($request);
@@ -57,6 +66,10 @@ class ProjectController extends AbstractController
      */
     public function show(Project $project): Response
     {
+        if (!$this->getUser()) {
+            return $this->redirectToRoute('root');
+        }
+
         $category = new Category();
         $formCategory = $this->createForm(CategoryType::class, $category, [
             'action' => $this->generateUrl('category_new'),
@@ -73,6 +86,10 @@ class ProjectController extends AbstractController
      */
     public function edit(Request $request, Project $project): Response
     {
+        if (!$this->getUser()) {
+            return $this->redirectToRoute('root');
+        }
+
         $form = $this->createForm(ProjectType::class, $project);
         $form->handleRequest($request);
 
@@ -96,6 +113,10 @@ class ProjectController extends AbstractController
      */
     public function delete(Request $request, Project $project): Response
     {
+        if (!$this->getUser()) {
+            return $this->redirectToRoute('root');
+        }
+        
         if ($this->isCsrfTokenValid('delete'.$project->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($project);
